# specs/001-qr-attendance-payroll/contracts/api.yaml
openapi: 3.1.0
info:
  title: AttendanceApp API
  version: 1.0.0
  description: |
    REST API for the QR-based employee attendance and payroll system.
    All endpoints (except /auth/login) require a valid Bearer JWT.
    All communication MUST use HTTPS/TLS 1.2+.
servers:
  - url: https://attendanceapp-api.onrender.com/api
    description: Production (Render.com)
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Employee:
      type: object
      properties:
        id:          { type: string }
        name:        { type: string }
        email:       { type: string, format: email }
        role:        { type: string, enum: [employee, admin] }
        status:      { type: string, enum: [active, inactive] }
        workplaceId: { type: string }
        createdAt:   { type: string, format: date-time }
    AttendanceRecord:
      type: object
      properties:
        id:                   { type: string }
        employeeId:           { type: string }
        date:                 { type: string, format: date, example: "2026-02-13" }
        checkInAt:            { type: string, format: date-time }
        checkOutAt:           { type: string, format: date-time, nullable: true }
        status:               { type: string, enum: [open, closed, incomplete] }
        regularHours:         { type: number, nullable: true }
        overtimeHoursMorning: { type: number, nullable: true }
        overtimeHoursEvening: { type: number, nullable: true }
        isManuallyAdjusted:   { type: boolean }
        adjustmentNote:       { type: string, nullable: true }
    PaysheetEntry:
      type: object
      properties:
        employeeId:           { type: string }
        employeeName:         { type: string }
        date:                 { type: string, format: date }
        checkInAt:            { type: string, format: date-time }
        checkOutAt:           { type: string, format: date-time }
        regularHours:         { type: number }
        overtimeHoursMorning: { type: number }
        overtimeHoursEvening: { type: number }
        regularPay:           { type: number, example: 1000 }
        overtimePay:          { type: number }
        totalPay:             { type: number }
        isManuallyAdjusted:   { type: boolean }
        recordStatus:         { type: string, enum: [included, skipped_incomplete] }
    Paysheet:
      type: object
      properties:
        id:              { type: string }
        generatedBy:     { type: string }
        generatedAt:     { type: string, format: date-time }
        periodStart:     { type: string, format: date }
        periodEnd:       { type: string, format: date }
        entries:
          type: array
          items: { $ref: '#/components/schemas/PaysheetEntry' }
        totals:
          type: object
          properties:
            totalRegularPay:  { type: number }
            totalOvertimePay: { type: number }
            totalPayable:     { type: number }
            skippedDays:      { type: integer }
        status: { type: string, enum: [draft, processed] }
...
security:
  - BearerAuth: []
paths:
...
  /admin/employees:
    get:
      tags: [admin-employees]
      summary: List all employees
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
        - in: query
          name: workplaceId
          schema: { type: string }
      responses:
        '200':
          description: Employee list
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items: { $ref: '#/components/schemas/Employee' }
    post:
      tags: [admin-employees]
      summary: Create a new employee account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:        { type: string }
                email:       { type: string, format: email }
                password:    { type: string, minLength: 8 }
                role:        { type: string, enum: [employee, admin] }
                workplaceId: { type: string }
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: '#/components/schemas/Employee' }
        '409':
          description: Email already in use
  /admin/employees/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    put:
      tags: [admin-employees]
      summary: Update employee details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                role:        { type: string, enum: [employee, admin] }
                workplaceId: { type: string }
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee: { $ref: '#/components/schemas/Employee' }
    patch:
      tags: [admin-employees]
      summary: Change employee active/inactive status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, inactive] }
      responses:
        '200':
          description: Status updated

  # ─── ADMIN — ATTENDANCE ───────────────────────────────────────────────────────
  /admin/attendance:
    get:
      tags: [admin-attendance]
      summary: Query attendance records
      parameters:
        - in: query
          name: employeeId
          schema: { type: string }
        - in: query
          name: startDate
          required: true
          schema: { type: string, format: date }
        - in: query
          name: endDate
          required: true
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { type: string, enum: [open, closed, incomplete] }
      responses:
        '200':
          description: Attendance records matching query
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items: { $ref: '#/components/schemas/AttendanceRecord' }
                  totalCount: { type: integer }
...
  /admin/attendance/{id}/close:
    patch:
      tags: [admin-attendance]
      summary: Manually close an incomplete attendance record
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checkOutAt, adjustmentNote]
              properties:
                checkOutAt:     { type: string, format: date-time }
                adjustmentNote: { type: string, minLength: 10 }
      responses:
        '200':
          description: Record closed with pay calculation applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  record: { $ref: '#/components/schemas/AttendanceRecord' }

  # ─── ADMIN — WORKPLACE ────────────────────────────────────────────────────────
...
  /admin/paysheets/generate:
    post:
      tags: [admin-paysheets]
      summary: Generate a paysheet for selected employees and date range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [periodStart, periodEnd, employeeIds]
              properties:
                periodStart:  { type: string, format: date }
                periodEnd:    { type: string, format: date }
                employeeIds:
                  type: array
                  items: { type: string }
                  description: Employee IDs to include; send empty array for all employees
      responses:
        '201':
          description: Paysheet generated (draft status)
          content:
            application/json:
              schema:
                type: object
                properties:
                  paysheet: { $ref: '#/components/schemas/Paysheet' }
  /admin/paysheets/{id}:
    get:
      tags: [admin-paysheets]
      summary: Get a paysheet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  paysheet: { $ref: '#/components/schemas/Paysheet' }
    patch:
      tags: [admin-paysheets]
      summary: Mark paysheet as processed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [processed] }
      responses:
        '200':
          description: Status updated
  /admin/paysheets/{id}/export:
    get:
      tags: [admin-paysheets]
      summary: Export paysheet as PDF or CSV
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: format
          required: true
          schema: { type: string, enum: [pdf, csv] }
      responses:
        '200':
          description: File download
          content:
            application/pdf:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string, format: binary }